+++ PusherServer.Tests/AcceptanceTests/Trigger.cs
@@ -11,19 +11,30 @@ namespace PusherServer.Tests.AcceptanceTests
     [TestFixture]
     public class When_Triggering_an_Event_on_a_single_Channel
     {
+        IPusher _pusher;
+
         [TestFixtureSetUp]
         public void Setup()
         {
             PusherClient.Pusher.Trace.Listeners.Add(new ConsoleTraceListener(true));
-
+            _pusher = new Pusher(Config.AppId, Config.AppKey, Config.AppSecret, new PusherOptions()
+            {
+                Host = Config.Host
+            });
         }
 
         [Test]
         public void It_should_return_a_200_response()
         {
-            IPusher pusher = new Pusher(Config.AppId, Config.AppKey, Config.AppSecret);
-            ITriggerResult result = pusher.Trigger("my-channel", "my_event", new { hello = "world" });
+            ITriggerResult result = _pusher.Trigger("my-channel", "my_event", new { hello = "world" });
             Assert.AreEqual(HttpStatusCode.OK, result.StatusCode);
+        }
+
+        [Test]
+        public void it_should_expose_the_event_id()
+        {
+            ITriggerResult result = _pusher.Trigger("my-channel", "my_event", new { hello = "world" });
+            Assert.IsTrue(string.IsNullOrEmpty(result.EventIds["my-channel"]) == false);
         }
 
         [Test]
@@ -34,7 +45,9 @@ namespace PusherServer.Tests.AcceptanceTests
 
             bool eventReceived = false;
             AutoResetEvent reset = new AutoResetEvent(false);
+
             var client = new PusherClient.Pusher(Config.AppKey);
+            client.Host = Config.WebSocketHost;
             client.Connected += new PusherClient.ConnectedEventHandler(delegate(object sender)
             {
                 Debug.WriteLine("connected");
@@ -55,6 +68,9 @@ namespace PusherServer.Tests.AcceptanceTests
                 reset.Set();
             });
 
+            Debug.WriteLine("waiting for Subscribed");
+            reset.WaitOne(TimeSpan.FromSeconds(5));
+
             Debug.WriteLine("binding");
             channel.Bind(eventName, delegate(dynamic data)
             {
@@ -63,15 +79,11 @@ namespace PusherServer.Tests.AcceptanceTests
                 reset.Set();
             });
 
-            Debug.WriteLine("waiting to subscribe");
-            reset.WaitOne(TimeSpan.FromSeconds(5));
-
-            Debug.WriteLine("triggering");
-            IPusher pusher = new Pusher(Config.AppId, Config.AppKey, Config.AppSecret);
-            ITriggerResult result = pusher.Trigger(channelName, eventName, new { hello = "world" });
+            Debug.WriteLine("Bound. Triggering");
+            ITriggerResult result = _pusher.Trigger(channelName, eventName, new { hello = "world" });
 
             Debug.WriteLine("waiting for event to be received");
-            reset.WaitOne(TimeSpan.FromSeconds(5));
+            reset.WaitOne(TimeSpan.FromSeconds(10));
 
             Assert.IsTrue(eventReceived);
         }
@@ -82,19 +94,22 @@ namespace PusherServer.Tests.AcceptanceTests
             var eventJSON = File.ReadAllText("AcceptanceTests/percent-message.json");
             var message = new JavaScriptSerializer().Deserialize(eventJSON, typeof(object));
 
-            IPusher pusher = new Pusher(Config.AppId, Config.AppKey, Config.AppSecret);
-            ITriggerResult result = pusher.Trigger("my-channel", "my_event", message);
+            ITriggerResult result = _pusher.Trigger("my-channel", "my_event", message);
             Assert.AreEqual(HttpStatusCode.OK, result.StatusCode);
         }
+        
     }
 
     [TestFixture]
     public class When_Triggering_an_Event_on_a_multiple_Channels
     {
         [Test]
-        public void It_should_return_a_202_response()
+        public void It_should_return_a_200_response()
         {
-            IPusher pusher = new Pusher(Config.AppId, Config.AppKey, Config.AppSecret);
+            IPusher pusher = new Pusher(Config.AppId, Config.AppKey, Config.AppSecret, new PusherOptions()
+            {
+                Host = Config.Host
+            });
             ITriggerResult result = pusher.Trigger(new string[] { "my-channel-1", "my-channel-2" }, "my_event", new { hello = "world" });
             Assert.AreEqual(HttpStatusCode.OK, result.StatusCode);
         }
@@ -103,18 +118,43 @@ namespace PusherServer.Tests.AcceptanceTests
     [TestFixture]
     public class When_Triggering_an_Event_over_HTTPS
     {
+        IPusher _pusher = null;
+
         [TestFixtureSetUp]
         public void Setup()
         {
             PusherClient.Pusher.Trace.Listeners.Add(new ConsoleTraceListener(true));
+
+            _pusher = new Pusher(Config.AppId, Config.AppKey, Config.AppSecret, new PusherOptions() {
+                Encrypted = true,
+                Host = Config.Host
+            });
         }
 
         [Test]
-        public void It_should_return_a_202_response()
+        public void It_should_return_a_200_response()
         {
-            IPusher pusher = new Pusher(Config.AppId, Config.AppKey, Config.AppSecret, new PusherOptions() { Encrypted = true } );
-            ITriggerResult result = pusher.Trigger("my-channel", "my_event", new { hello = "world" });
+            ITriggerResult result = _pusher.Trigger("my-channel", "my_event", new { hello = "world" });
             Assert.AreEqual(HttpStatusCode.OK, result.StatusCode);
+        }
+
+        [Test]
+        public void It_should_expose_a_single_event_id_when_publishing_to_a_single_channel()
+        {
+            ITriggerResult result = _pusher.Trigger("ch1", "my_event", new { hello = "world" });
+            Assert.IsTrue(result.EventIds.ContainsKey("ch1"));
+            Assert.AreEqual(1, result.EventIds.Count);
+        }
+
+        [Test]
+        public void It_should_expose_a_multiple_event_ids_when_publishing_to_multiple_channels()
+        {
+            var channels = new string[]{"ch1", "ch2", "ch3"};
+            ITriggerResult result = _pusher.Trigger(channels, "my_event", new { hello = "world" });
+            Assert.IsTrue(result.EventIds.ContainsKey("ch1"));
+            Assert.IsTrue(result.EventIds.ContainsKey("ch2"));
+            Assert.IsTrue(result.EventIds.ContainsKey("ch3"));
+            Assert.AreEqual(channels.Length, result.EventIds.Count);
         }
     }
 }