+++ PusherServer.Tests/UnitTests/PusherOptions.cs
@@ -15,6 +15,37 @@ namespace PusherServer.Tests.UnitTests
         }
 
         [Test]
+        public void Host_defaults_to_api_pusherapp_com()
+        {
+            var options = new PusherOptions();
+            Assert.AreEqual("api.pusherapp.com", options.Host);
+        }
+
+        [Test]
+        [ExpectedException(typeof(FormatException))]
+        public void https_scheme_is_not_allowed_when_setting_host()
+        {
+            var httpsOptions = new PusherOptions();
+            httpsOptions.Host = "https://api.pusherapp.com";
+        }
+
+        [Test]
+        [ExpectedException(typeof(FormatException))]
+        public void http_scheme_is_not_allowed_when_setting_host()
+        {
+            var httpsOptions = new PusherOptions();
+            httpsOptions.Host = "http://api.pusherapp.com";
+        }
+
+        [Test]
+        [ExpectedException(typeof(FormatException))]
+        public void ftp_scheme_is_not_allowed_when_setting_host()
+        {
+            var httpsOptions = new PusherOptions();
+            httpsOptions.Host = "ftp://api.pusherapp.com";
+        }
+        
+        [Test]
         public void Port_defaults_to_80()
         {
             var options = new PusherOptions();