+++ PusherServer/ITriggerResult.cs
@@ -1,8 +1,9 @@
-﻿using System.Net;
+﻿using System.Collections.Generic;
 
 namespace PusherServer
 {
     public interface ITriggerResult: IRequestResult
     {
+        IDictionary<string, string> EventIds { get; }
     }
 }