+++ PusherServer/Pusher.cs
@@ -4,6 +4,7 @@ using RestSharp;
 using RestSharp.Serializers;
 using System.Text.RegularExpressions;
 using System.Reflection;
+using System.Diagnostics;
 
 namespace PusherServer
 {
@@ -13,7 +14,7 @@ namespace PusherServer
     /// </summary>
     public class Pusher : IPusher
     {
-        private static string DEFAULT_REST_API_HOST = "api.pusherapp.com";
+        
 
         private string _appId;
         private string _appKey;
@@ -230,7 +231,21 @@ namespace PusherServer
 
             var request = CreateAuthenticatedRequest(Method.POST, "/events", null, requestBody);
 
+            Debug.WriteLine(string.Format("Method: {1}{0}Host: {2}{0}Resource: {3}{0}Parameters:{4}",
+                Environment.NewLine,
+                request.Method,
+                _options.RestClient.BaseUrl,
+                request.Resource, 
+                string.Join(",", Array.ConvertAll(request.Parameters.ConvertAll(p => p.Name + "=" + p.Value).ToArray(), i => i.ToString()))
+            ));
+
             IRestResponse response = _options.RestClient.Execute(request);
+
+            Debug.WriteLine(string.Format("Response{0}StatusCode: {1}{0}Body: {2}",
+                Environment.NewLine,
+                response.StatusCode,
+                response.Content));
+
             return response;
         }
 
@@ -310,7 +325,7 @@ namespace PusherServer
         private Uri GetBaseUrl(IPusherOptions _options)
         {
             string baseUrl = (_options.Encrypted ? "https" : "http") + "://" +
-                DEFAULT_REST_API_HOST +
+                _options.Host +
                 (_options.Port == 80 ? "" : ":" + _options.Port);
             return new Uri( baseUrl );
         }