+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.0.0
+
+* [CHANGED] `Trigger` calls that result in a non 200 response from the Pusher HTTP API now result in a `TriggerResponseException` being thrown.
+			This is a BREAKING CHANGE as previously you could inspect the `ITriggerResult.StatusCode` to detect a failed request.
+* [ADDED]	When triggering events against a Pusher cluster that supports Event Buffer functionality the IDs of the triggered events
+			can be retrieved via `ITriggerResult.EventIds`
+* [ADDED]	The Pusher HTTP API Host can now be set via `PusherOptions.Host`
+
 ## 2.1.1
 
 * [FIXED] Channel name and socket_id values are validated for `Pusher.Trigger`